Day-8/requirements.txt
flask
flask-cors
paddleocr
paddlepaddle
opencv-python-headless
numpy
Pillow
PyMuPDF
werkzeug